+++ babel.config.js
@@ -1,5 +1,3 @@
-'use strict';
-
 module.exports = {
   presets: ['@babel/preset-env', '@babel/flow'],
 };