+++ src/__tests__/parsers/typescript_parser.test.js
@@ -3,11 +3,9 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow
  */
 
-const {parse} = require('../../parsers/babylon_parser');
-const {parserTests} = require('../../../fixtures/parser_tests');
+import {parse} from '../../parsers/typescript_parser';
+import {parserTests} from '../../../fixtures/parser_tests';
 
 parserTests(parse);