+++ src/__tests__/project_workspace.test.js
@@ -7,19 +7,26 @@
  * @flow
  */
 
-'use strict';
-
 import ProjectWorkspace from '../project_workspace';
 
 describe('setup', () => {
   it('sets itself up fom the constructor', () => {
-    const workspace = new ProjectWorkspace(
-      'root_path',
-      'path_to_jest',
-      'path_to_config',
-      1000,
-    );
+    const workspace = new ProjectWorkspace('root_path', 'path_to_jest', 'path_to_config', 1000);
     expect(workspace.rootPath).toEqual('root_path');
     expect(workspace.pathToJest).toEqual('path_to_jest');
+  });
+  it('can safe guard outputFileSuffix', () => {
+    const suffixStrings = [
+      [undefined, undefined],
+      ['with_underscore', 'with_underscore'],
+      ['camelCase', 'camelcase'],
+      ['upperCase', 'uppercase'],
+      ['with Space', 'with_space'],
+      ['WITH?special ? character\n', 'with_special___character_'],
+    ];
+    suffixStrings.forEach(pair => {
+      const workspace = new ProjectWorkspace('root_path', 'path_to_jest', 'path_to_config', 1000, pair[0]);
+      expect(workspace.outputFileSuffix).toEqual(pair[1]);
+    });
   });
 });